new update

+++ src/Components/delete/delete.jsx
@@ -4,29 +4,43 @@ import './delete.css';
 import axios from 'axios';
 
 const Delete = ({ categories }) => {
-    const [selectedCategoryId, setSelectedCategoryId] = useState("All");
-    const [showConfirm, setShowConfirm] = useState(false);
-    const [showErrror, setShowError] = useState(false);
     
     const userInfo = localStorage.getItem('userInfo');
     const {data} = JSON.parse(userInfo)
-  
-  
-    useEffect(() => {
-        console.log(data);
-        console.log(categories)
-    }, []);
+
+    const [selectedCategoryId, setSelectedCategoryId] = useState("All");
+    const [showConfirm, setShowConfirm] = useState(false);
+    const [showErrror, setShowError] = useState(false);
+    const [updetedCategory, setUpdetedCategory] = useState({})
     
 
 
     const handleSelect = async (event) => {
         const { value } = event.target;
-        setSelectedCategoryId(value);
+        setSelectedCategoryId(value);   
+        const selectedCategory = categories.find(category => category._id === value);
+        const selectedCategoryByUserId = selectedCategory.userId;
+        const selectedRole = selectedCategory.subscription.find(sub => sub.role == 'teacher')
+        console.log(selectedRole);
+        setUpdetedCategory({
+            theCategory: selectedCategory.courseName,
+            role: selectedRole.role
+        })
+        console.log(updetedCategory);
+        console.log();
+        console.log(selectedCategoryByUserId);
+        console.log(data.user._id);
+        if(selectedCategoryByUserId !== data.user._id){
+            setShowError(true)
+            return
+        }
+
         const theValue = value
         if (value !== "All") {
             try{
-                console.log(theValue);
+                console.log(updetedCategory);
                 await axios.delete(`http://localhost:3000/courses/${theValue}`, {withCredentials: true})
+                console.log('put');
                 setShowConfirm(true);
             }catch(error){
                 setShowError(true)
@@ -36,8 +50,9 @@ const Delete = ({ categories }) => {
     };
 
     const handleConfirmDelete = async () => {
+        console.log(updetedCategory);
         try{
-        await axios.delete(`http://localhost:3000/courses/${selectedCategoryId}`, {withCredentials: true})
+        await axios.delete(`http://localhost:3000/courses/${selectedCategoryId}`, updetedCategory, {withCredentials: true})
         }
         catch (error) {
             console.error('Error delete course:', error);
